poc/receive_email/requirements.txt
apscheduler>=3.10.0
python-dotenv>=1.0.0
pdfplumber>=0.9.0
python-docx>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
redis>=5.0.0
